Extract QuestionUi sealed class to its own class
